+++ test/multidimensional.jl
@@ -35,6 +35,7 @@ function test_nufft_type1(
         σ = 1.25,
         block_size = NonuniformFFTs.default_block_size(Ns, CPU()),
         sort_points = False(),
+        kws...,
     ) where {T <: Number}
     Tr = real(T)
     ks = map(N -> fftfreq(N, Tr(N)), Ns)
@@ -62,7 +63,7 @@ function test_nufft_type1(
 
     # Compute NUFFT
     ûs = Array{Complex{Tr}}(undef, map(length, ks))
-    plan_nufft = @inferred PlanNUFFT(T, Ns, m; σ, kernel, block_size, sort_points)
+    plan_nufft = @inferred PlanNUFFT(T, Ns, m; σ, kernel, block_size, sort_points, kws...)
     NonuniformFFTs.set_points!(plan_nufft, xp)
     NonuniformFFTs.exec_type1!(ûs, plan_nufft, vp)
 
@@ -163,6 +164,9 @@ end
     @testset "Sort points" begin
         @testset "Type 1 NUFFT" test_nufft_type1(T, Ns; sort_points = True())
         @testset "Type 2 NUFFT" test_nufft_type2(T, Ns; sort_points = True())
+    end
+    @testset "Using atomics" begin
+        test_nufft_type1(T, Ns; use_atomics = true)
     end
     @testset "Non-multiple of block size" begin
         # This gives an oversampled grid size Ñs = (80, 80) (when σ = 2.0).